# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "im在线通信系统前后台接口文档",
        "title": "Im在线通信系统接口文档",
        "contact": {},
        "version": "1.0 版本"
    },
    "host": "localhost:8080",
    "paths": {
        "/common/email": {
            "get": {
                "description": "发送邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端通用接口"
                ],
                "summary": "发送邮箱验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/common/test": {
            "get": {
                "description": "测试接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端通用接口"
                ],
                "summary": "测试接口",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/common/upload": {
            "post": {
                "description": "文件上传接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端通用接口"
                ],
                "summary": "文件上传接口",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"ok\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/front/room/applyJoinRoom": {
            "post": {
                "description": "申请加入房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端房间接口"
                ],
                "summary": "申请加入房间",
                "parameters": [
                    {
                        "description": "用户申请信息",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RoomApply"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        },
        "/front/room/createRoom": {
            "post": {
                "description": "创建房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端房间接口"
                ],
                "summary": "创建房间",
                "parameters": [
                    {
                        "description": "房间信息",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        },
        "/front/room/getRoomApply": {
            "get": {
                "description": "获取房间的申请信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端房间接口"
                ],
                "summary": "获取房间的申请信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "房主id",
                        "name": "master_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-models_RoomApply"
                        }
                    }
                }
            }
        },
        "/front/user/addFriend": {
            "post": {
                "description": "请求添加好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "请求添加好友",
                "parameters": [
                    {
                        "description": "处理申请的参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        },
        "/front/user/deleteFriend": {
            "post": {
                "description": "删除好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "删除好友",
                "parameters": [
                    {
                        "description": "处理申请的参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        },
        "/front/user/getUserFriendApply": {
            "get": {
                "description": "获取用户好友申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "获取用户好友申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Result"
                        }
                    }
                }
            }
        },
        "/front/user/getUserFriends": {
            "get": {
                "description": "获取用户好友",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "获取用户好友",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-array_models_Users"
                        }
                    }
                }
            }
        },
        "/front/user/handleApplication": {
            "post": {
                "description": "处理用户好友申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "处理用户好友申请",
                "parameters": [
                    {
                        "description": "处理申请的参数",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserApplyDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        },
        "/front/user/login": {
            "post": {
                "description": "user login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "用户登录接口",
                "parameters": [
                    {
                        "description": "Id",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-models_Users"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/front/user/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "前端用户接口"
                ],
                "summary": "用户注册接口",
                "parameters": [
                    {
                        "description": "注册用户的信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.RT-string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.RT-array_models_Users": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Users"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.RT-models_RoomApply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.RoomApply"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.RT-models_Users": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/models.Users"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.RT-string": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "common.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "errs": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "dto.UserApplyDto": {
            "type": "object",
            "required": [
                "is_accept",
                "user_id"
            ],
            "properties": {
                "friend_id": {
                    "type": "string"
                },
                "is_accept": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.UserLoginDto": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.Room": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "master_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "models.RoomApply": {
            "type": "object",
            "required": [
                "master_id",
                "user_id"
            ],
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "master_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.UserFriend": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "friend_id": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "sex"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "create_time": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "synopsis": {
                    "type": "string"
                },
                "update_time": {
                    "type": "string"
                }
            }
        }
    }
}